+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 Check our main [developer changelog](https://developer.paddle.com/?utm_source=dx&utm_medium=paddle-php-sdk) for information about changes to the Paddle Billing platform, the Paddle API, and other developer tools.
 
+## [unreleased] - 2024-10-01
+
+### Added
+
+- Added `traffic_source` property to `NotificationSetting` entity
+- Support notification settings `traffic_source` filter
+
 # [1.3.1] - 2024-09-30
 
 ### Fixed